requirements.txt
biopython>=1.79
matplotlib>=3.5.0
pandas>=1.3.0
numpy>=1.21.0
seaborn>=0.11.0
ete3>=3.1.2
progress>=1.6

